+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The following changes applies since 5.0.0-beta1, and do not affect 4.x
 - Fix potential UB in bitstring_to_u64 with large input and many ignored bits
 - Fix constructed objects parsing with indefinite length (do not include EOC)
 - Constructed objects: use `InvalidTag` everywhere if tag is not expected
+- Add new error `StringInvalidCharset` and update string parsing methods
 
 ## 5.0.0-beta2